# xcube_hub/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: info@brockmann-consult.de
  description: Restful API for handling xcube Services
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: xcube Generation API
  version: 2.0.0
servers:
- url: https://xcube-gen.brockmann-consult.de/api/v2
- url: https://stage.xcube-gen.brockmann-consult.de/api/v2
- url: https://catehub.climate.esa.int/api/v2
- url: https://stage.catehub.climate.esa.int/api/v2
...
paths:
...
  /users:
    get:
      description: |
        Get all users by service name
      operationId: get_users
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiUsersResponse'
          description: List of users
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Api Error
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: No users found
      security:
      - oAuthorization:
        - read:users
      summary: Get all users by service name
      tags:
      - users
      x-openapi-router-controller: xcube_hub.controllers.users
  /users/{user_id}:
    delete:
      description: |
        Remove user from a service
      operationId: delete_user_by_user_id
      parameters:
      - description: User ID
        explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: User deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Api Error
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: User not found
      security:
      - oAuthorization:
        - delete:users
      summary: Delete user
      tags:
      - users
      x-openapi-router-controller: xcube_hub.controllers.users
    get:
      description: |
        Get user info
      operationId: get_user_by_user_id
      parameters:
      - description: User ID
        explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiUserResponse'
          description: User information
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Api Error
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: User not found in the service
      security:
      - oAuthorization:
        - read:users
      summary: Get users
      tags:
      - users
      x-openapi-router-controller: xcube_hub.controllers.users
    patch:
      description: |
        Update users in a service
      operationId: update_user_by_user_id
      parameters:
      - description: User ID
        explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: User information
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiUserResponse'
          description: User updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Api Error
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: User not found
      security:
      - oAuthorization:
        - update:users
      summary: Update user
      tags:
      - users
      x-openapi-router-controller: xcube_hub.controllers.users
    put:
      description: |
        Add user
      operationId: add_user
      parameters:
      - description: User ID
        explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: User information
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiUserResponse'
          description: User added
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: Api Error
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
          description: User not found in the service
      security:
      - oAuthorization:
        - create:users
      summary: Add user
      tags:
      - users
      x-openapi-router-controller: xcube_hub.controllers.users
components:
  schemas:
    ApiResponse:
      properties:
        message:
          type: string
        output:
          type: object
      type: object
    ApiErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
...
    ApiUserResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
...
    ApiUsersResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
...
    User:
      example:
        user_metadata: '{}'
        updated_at: 2000-01-23T04:56:07.000+00:00
        user_id: user_id
        nickname: nickname
        created_at: 2000-01-23T04:56:07.000+00:00
        connection: connection
        given_name: given_name
        family_name: family_name
        email: email
        username: username
        app_metadata: '{}'
      properties:
        user_id:
          type: string
        email:
          type: string
        username:
          type: string
        nickname:
          type: string
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
        given_name:
          type: string
        family_name:
          type: string
        app_metadata:
          type: object
        user_metadata:
          type: object
        connection:
          type: string
      type: object
...
  securitySchemes:
    oAuthorization:
      description: This API uses Oauth 2 with the implicit grant flow.
      flows:
        clientCredentials:
          scopes:
            read_cubegen: Read cubegen status and result
            read_punits: Allow reading the punits level of a user
            put_punits: Allow adding punits to and remove punits from a user
            read_callbacks: Allow reading callbacks
            put_callbacks: Allow reading callbacks
            delete_callbacks: Allow reading callbacks
          tokenUrl: https://edc.eu.auth0.com/oauth/token
        implicit:
          authorizationUrl: https://edc.eu.auth0.com/authorize
          scopes:
            read_cubegen: Read cubegen status and result
            submit_cubegen: Submit cubegens to the K8s cluster
            read_punits: Allow reading the punits level of a user
            read_callbacks: Allow reading callbacks
      type: oauth2
      x-tokenInfoFunc: xcube_hub.controllers.authorization.check_oAuthorization
      x-scopeValidateFunc: xcube_hub.controllers.authorization.validate_scope_oAuthorization